Ensure imports are rewritten before setting up context
